# Conflicts:
#	web-app/src/main/kotlin/com/respiroc/webapp/filter/TenantIdFilter.kt

+++ util/src/main/resources/db/migration/V0_0_5__role_permissions_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS role_permissions
 (
-    id            BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    role_id       BIGINT NOT NULL,
-    permission_id BIGINT NOT NULL,
+    id            INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    role_id       INTEGER NOT NULL,
+    permission_id INTEGER NOT NULL,
 
     CONSTRAINT fk__role_permissions__role FOREIGN KEY (role_id) REFERENCES roles (id) ON DELETE CASCADE,
     CONSTRAINT fk__role_permissions__permission FOREIGN KEY (permission_id) REFERENCES permissions (id) ON DELETE CASCADE,
