+++ util/src/main/resources/db/migration/V0_0_6__user_roles_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS user_roles
 (
-    id      BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    role_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    id      INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    role_id INTEGER NOT NULL,
+    user_id INTEGER NOT NULL,
 
     CONSTRAINT fk__user_roles__role FOREIGN KEY (role_id) REFERENCES roles (id) ON DELETE CASCADE,
     CONSTRAINT fk__user_roles__user FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE,
