+++ util/src/main/resources/db/migration/V0_1_1__user_sessions_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS user_sessions
 (
-    id               BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    user_id          BIGINT    NOT NULL,
+    id               INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    user_id          INTEGER    NOT NULL,
     token            TEXT      NOT NULL UNIQUE,
     token_issue_at   TIMESTAMP NOT NULL,
     token_expire_at  TIMESTAMP NOT NULL,
